LPS-154708 Move BundleStartStopLogger to portal-bootstrap
