+++ modules/core/portal-bootstrap/src/main/java/com/liferay/portal/bootstrap/log/BundleStartStopLogger.java
@@ -12,24 +12,55 @@
  * details.
  */
 
-package com.liferay.portal.equinox.log.bridge.internal;
+package com.liferay.portal.bootstrap.log;
+
+import com.liferay.petra.string.StringBundler;
+import com.liferay.portal.kernel.log.Log;
+import com.liferay.portal.kernel.log.LogFactoryUtil;
 
 import java.util.concurrent.atomic.AtomicBoolean;
 
 import org.osgi.framework.Bundle;
+import org.osgi.framework.BundleContext;
 import org.osgi.framework.BundleEvent;
+import org.osgi.framework.InvalidSyntaxException;
+import org.osgi.framework.ServiceReference;
 import org.osgi.framework.SynchronousBundleListener;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import org.osgi.util.tracker.ServiceTracker;
 
 /**
  * @author Raymond Augé
  */
 public class BundleStartStopLogger implements SynchronousBundleListener {
 
-	public BundleStartStopLogger(AtomicBoolean portalStarted) {
-		_portalStarted = portalStarted;
+	public BundleStartStopLogger(BundleContext bundleContext)
+		throws InvalidSyntaxException {
+
+		ServiceTracker<Object, Void> serviceTracker =
+			new ServiceTracker<Object, Void>(
+				bundleContext,
+				bundleContext.createFilter(
+					StringBundler.concat(
+						"(&(objectClass=",
+						"com.liferay.portal.kernel.module.framework.",
+						"ModuleServiceLifecycle)",
+						"(module.service.lifecycle=portal.initialized))")),
+				null) {
+
+				@Override
+				public Void addingService(
+					ServiceReference<Object> serviceReference) {
+
+					_portalStarted.set(true);
+
+					close();
+
+					return null;
+				}
+
+			};
+
+		serviceTracker.open();
 	}
 
 	@Override
@@ -37,32 +68,32 @@ public class BundleStartStopLogger implements SynchronousBundleListener {
 		Bundle bundle = bundleEvent.getBundle();
 
 		if (bundle.getSymbolicName() == null) {
-			_log.error("{} has a null symbolic name", bundle.getLocation());
+			_log.error(bundle.getLocation() + " has a null symbolic name");
 		}
 
 		if (_portalStarted.get()) {
 			if (_log.isInfoEnabled()) {
 				if (bundleEvent.getType() == BundleEvent.STARTED) {
-					_log.info("STARTED {}", bundle);
+					_log.info("STARTED " + bundle);
 				}
 				else if (bundleEvent.getType() == BundleEvent.STOPPED) {
-					_log.info("STOPPED {}", bundle);
+					_log.info("STOPPED " + bundle);
 				}
 			}
 		}
 		else if (_log.isDebugEnabled()) {
 			if (bundleEvent.getType() == BundleEvent.STARTED) {
-				_log.debug("STARTED {}", bundle);
+				_log.debug("STARTED " + bundle);
 			}
 			else if (bundleEvent.getType() == BundleEvent.STOPPED) {
-				_log.debug("STOPPED {}", bundle);
+				_log.debug("STOPPED " + bundle);
 			}
 		}
 	}
 
-	private static final Logger _log = LoggerFactory.getLogger(
+	private static final Log _log = LogFactoryUtil.getLog(
 		BundleStartStopLogger.class);
 
-	private final AtomicBoolean _portalStarted;
+	private final AtomicBoolean _portalStarted = new AtomicBoolean();
 
 }
