+++ modules/core/portal-equinox-log-bridge/src/main/java/com/liferay/portal/equinox/log/bridge/internal/PortalHookConfigurator.java
@@ -14,10 +14,6 @@
 
 package com.liferay.portal.equinox.log.bridge.internal;
 
-import com.liferay.petra.string.StringBundler;
-
-import java.util.concurrent.atomic.AtomicBoolean;
-
 import org.eclipse.equinox.log.ExtendedLogReaderService;
 import org.eclipse.osgi.internal.hookregistry.ActivatorHookFactory;
 import org.eclipse.osgi.internal.hookregistry.HookConfigurator;
@@ -25,9 +21,7 @@ import org.eclipse.osgi.internal.hookregistry.HookRegistry;
 
 import org.osgi.framework.BundleActivator;
 import org.osgi.framework.BundleContext;
-import org.osgi.framework.InvalidSyntaxException;
 import org.osgi.framework.ServiceReference;
-import org.osgi.util.tracker.ServiceTracker;
 
 /**
  * @author Raymond Augé
@@ -35,10 +29,6 @@ import org.osgi.util.tracker.ServiceTracker;
  */
 public class PortalHookConfigurator
 	implements ActivatorHookFactory, BundleActivator, HookConfigurator {
-
-	public PortalHookConfigurator() {
-		_bundleStartStopLogger = new BundleStartStopLogger(_portalStarted);
-	}
 
 	@Override
 	public void addHooks(HookRegistry hookRegistry) {
@@ -51,9 +41,7 @@ public class PortalHookConfigurator
 	}
 
 	@Override
-	public void start(BundleContext bundleContext)
-		throws InvalidSyntaxException {
-
+	public void start(BundleContext bundleContext) {
 		ServiceReference<ExtendedLogReaderService> serviceReference =
 			bundleContext.getServiceReference(ExtendedLogReaderService.class);
 
@@ -64,41 +52,10 @@ public class PortalHookConfigurator
 			extendedLogReaderService.addLogListener(
 				_portalSynchronousLogListener);
 		}
-
-		bundleContext.addBundleListener(_bundleStartStopLogger);
-
-		_serviceTracker = new ServiceTracker<Object, Void>(
-			bundleContext,
-			bundleContext.createFilter(
-				StringBundler.concat(
-					"(&(objectClass=",
-					"com.liferay.portal.kernel.module.framework.",
-					"ModuleServiceLifecycle)",
-					"(module.service.lifecycle=portal.initialized))")),
-			null) {
-
-			@Override
-			public Void addingService(
-				ServiceReference<Object> serviceReference) {
-
-				_portalStarted.set(true);
-
-				close();
-
-				return null;
-			}
-
-		};
-
-		_serviceTracker.open();
 	}
 
 	@Override
 	public void stop(BundleContext bundleContext) {
-		_serviceTracker.close();
-
-		bundleContext.removeBundleListener(_bundleStartStopLogger);
-
 		ExtendedLogReaderService extendedLogReaderService =
 			bundleContext.getService(
 				bundleContext.getServiceReference(
@@ -108,10 +65,7 @@ public class PortalHookConfigurator
 			_portalSynchronousLogListener);
 	}
 
-	private final BundleStartStopLogger _bundleStartStopLogger;
-	private final AtomicBoolean _portalStarted = new AtomicBoolean();
 	private final PortalSynchronousLogListener _portalSynchronousLogListener =
 		new PortalSynchronousLogListener();
-	private ServiceTracker<Object, Void> _serviceTracker;
 
 }
